refactor: 2.0 版本重构
新增：
1. 新增EventBus
2. 新增 邮箱、手机、用户名 + 密码登陆
3. 新增 邮箱 + 验证码 登录
4. 新增 手机号 + 验证码 登录
5. 新增个人修改手机号 前置身份校验
6. 新增个人修改邮箱 前置身份校验
7. 新增个人中心忘记密码 前置身份校验
8. 新增使用ResultWrapper 代替 ResultVo 返回数据

优化：
1. 优化接口加密
2. 优化CrudService实现
3. 变更自动日志记录模式到自定义记录
4. 优化导出Excel模式为认证模式
5. 优化用户Redis缓存
6. 移除Shiro，变更使用SpringSecurity 作为权限验证

升级/修复：
1. 修复 SecurityCache 缓存穿透BUG
2. 升级 mybatis-plus 版本至 3.5.2 修复若干CVE
3. 升级 springboot 版本至 2.5.6 修复若干CVE
4. 升级 mysql版本至8.0.28  修复 CVE-2022-21363
5. 升级 guava 版本至 30.0.android 修复 修复 CVE-2020-8908
6. 升级 bouncycastle 版本至 1.69 修复 Cxa9261daf-3755
7. 升级 protobuf-java 版本至 3.18.2 修复 CVE-2021-22569
8. 升级 logback 版本至 修复 CVE-2021-42550

+++ opsli-modulars/opsli-modulars-generator/src/main/java/org/opsli/modulars/generator/template/service/impl/GenTemplateServiceImpl.java
@@ -51,7 +51,7 @@ import java.util.List;
 /**
  * 代码模板 Service Impl
  *
- * @author 周鹏程
+ * @author Parker
  * @date 2021-05-27 14:33:49
  */
 @Service
@@ -191,7 +191,7 @@ public class GenTemplateServiceImpl extends CrudServiceImpl<GenTemplateMapper, G
         QueryWrapper<GenTemplate> queryWrapper = new QueryWrapper<>();
         queryWrapper.eq("table_type", base.getTableType());
         queryWrapper.notIn(FieldUtil.humpToUnderline(MyBatisConstants.FIELD_ID), id);
-        int count = this.count(queryWrapper);
+        long count = this.count(queryWrapper);
         if(count == 0){
             // 代码模板同一表类型下，至少保障有一个模板
             throw new GeneratorException(GeneratorMsg.EXCEPTION_TEMPLATE_AT_LEAST_ONE);
